Extract common methods to the abstract class
